#9, #10, relaciono productos con supermercado
